Refactor runner implementations into dedicated modules
