+++ projects/04-llm-adapter-shadow/src/llm_adapter/runner_shared.py
@@ -1,0 +1,67 @@
+"""Shared helpers for runner modules."""
+
+from __future__ import annotations
+
+from pathlib import Path
+
+from .errors import FatalError, RateLimitError, RetryableError, SkipError
+from .observability import EventLogger, JsonlLogger
+
+MetricsPath = str | Path | None
+
+
+def resolve_event_logger(
+    logger: EventLogger | None,
+    metrics_path: MetricsPath,
+) -> tuple[EventLogger | None, str | None]:
+    """Resolve the event logger and materialized metrics path."""
+    metrics_path_str = None if metrics_path is None else str(Path(metrics_path))
+    if metrics_path_str is None:
+        return None, None
+    if logger is not None:
+        return logger, metrics_path_str
+    return JsonlLogger(metrics_path_str), metrics_path_str
+
+
+def error_family(error: Exception | None) -> str | None:
+    if error is None:
+        return None
+    if isinstance(error, RateLimitError):
+        return "rate_limit"
+    if isinstance(error, SkipError):
+        return "skip"
+    if isinstance(error, FatalError):
+        return "fatal"
+    if isinstance(error, RetryableError):
+        return "retryable"
+    return "unknown"
+
+
+def estimate_cost(provider: object, tokens_in: int, tokens_out: int) -> float:
+    estimator = getattr(provider, "estimate_cost", None)
+    if callable(estimator):
+        try:
+            return float(estimator(tokens_in, tokens_out))
+        except Exception:  # pragma: no cover - defensive guard
+            return 0.0
+    return 0.0
+
+
+def provider_model(provider: object, *, allow_private: bool = False) -> str | None:
+    attrs = ["model"]
+    if allow_private:
+        attrs.append("_model")
+    for attr in attrs:
+        value = getattr(provider, attr, None)
+        if isinstance(value, str) and value:
+            return value
+    return None
+
+
+__all__ = [
+    "MetricsPath",
+    "resolve_event_logger",
+    "error_family",
+    "estimate_cost",
+    "provider_model",
+]
